[trace] break down otlp support with build tags
