+++ pkg/trace/api/otlp.go
@@ -3,6 +3,9 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
+//go:build otlp
+// +build otlp
+
 package api
 
 import (
