+++ pkg/trace/config/config_nootlp.go
@@ -1,0 +1,12 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+//go:build !otlp
+// +build !otlp
+
+package config
+
+// OTLP holds the configuration for the OpenTelemetry receiver.
+type OTLP struct{}
